Python/weather-app/requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==3.1.0
click==8.1.2
colorama==0.4.4
distlib==0.3.4
filelock==3.6.0
Flask==2.0.3
Flask-Login==0.6.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.1
jsonify==0.5
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
pipenv==2022.1.8
platformdirs==2.5.1
pycodestyle==2.8.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
toml==0.10.2
urllib3==2.0.2
virtualenv==20.13.3
virtualenv-clone==0.5.7
Werkzeug==2.1.1
